Implement `size_hint` and `ExactSizedIterator` for `DecimalArray` (#1506)
* Implement size_hint and ExactSizedIterator for DecimalArray

* clippy

* Apply suggestions from code review


* Move DecimalIter to iterator.rs

* Bring back doc fixes
